+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This library adheres to [Semantic Versioning](https://semver.org/) and
 [Keep a CHANGELOG](https://keepachangelog.com/en/1.0.0/).
 
+## 0.4.0
+
+- Change default output to JSON format for CLI tool.
+- Moved the current dump format to `--format dump`.
+- Added `--format text` option for plain text output.
+
 ## 0.3.0
 
 - Unwrap text in the changelog file for better readability.